+++ src/js/components/Navigation/index.ts
@@ -1,3 +1,4 @@
 import Navigation from './Navigation';
 
 export default Navigation;
+export * from './Navigation';
