+++ src/js/components/index.ts
@@ -1,0 +1,50 @@
+export * from './Album';
+export { default as Album } from './Album';
+export * from './AlbumTrack';
+export { default as AlbumTrack } from './AlbumTrack';
+export * from './AlbumTracks';
+export { default as AlbumTracks } from './AlbumTracks';
+export * from './Albums';
+export { default as Albums } from './Albums';
+export * from './Artist';
+export { default as Artist } from './Artist';
+export * from './Artists';
+export { default as Artists } from './Artists';
+export * from './Aside';
+export { default as Aside } from './Aside';
+export * from './Button';
+export { default as Button } from './Button';
+export * from './Categories';
+export { default as Categories } from './Categories';
+export * from './Category';
+export { default as Category } from './Category';
+export * from './Content';
+export { default as Content } from './Content';
+export * from './Error';
+export { default as Error } from './Error';
+export * from './Footer';
+export { default as Footer } from './Footer';
+export * from './Header';
+export { default as Header } from './Header';
+export * from './Image';
+export { default as Image } from './Image';
+export { default as Loader } from './Loader';
+export * from './Main';
+export { default as Main } from './Main';
+export * from './Media';
+export { default as Media } from './Media';
+export * from './Navigation';
+export { default as Navigation } from './Navigation';
+export { default as Player } from './Player';
+export * from './Playlist';
+export { default as Playlist } from './Playlist';
+export * from './PlaylistTrack';
+export { default as PlaylistTrack } from './PlaylistTrack';
+export * from './PlaylistTracks';
+export { default as PlaylistTracks } from './PlaylistTracks';
+export * from './Playlists';
+export { default as Playlists } from './Playlists';
+export * from './Search';
+export { default as Search } from './Search';
+export * from './User';
+export { default as User } from './User';
