+++ src/js/layouts/Default/Default.tsx
@@ -1,15 +1,17 @@
 import React, { Suspense } from 'react';
 import { Outlet } from 'react-router-dom';
+import {
+  Aside,
+  Content,
+  Error,
+  Header,
+  Loader,
+  Main,
+  Navigation,
+  Player,
+  Search,
+} from '../../components';
 import { useMePlaylists } from '../../hooks/me';
-import Aside from '../../components/Aside';
-import Content from '../../components/Content';
-import Error from '../../components/Error';
-import Header from '../../components/Header';
-import Loader from '../../components/Loader';
-import Main from '../../components/Main';
-import Navigation from '../../components/Navigation';
-import Player from '../../components/Player';
-import Search from '../../components/Search';
 import styles from './Default.module.css';
 
 function DefaultLayout(): JSX.Element {
