+++ src/js/pages/Browse/FeaturedPlaylists/index.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
-import Error from '../../../components/Error';
-import Loader from '../../../components/Loader';
-import Playlists from '../../../components/Playlists';
+import { Error, Loader, Playlists } from '../../../components';
 import { useBrowseFeaturedPlaylists } from '../../../hooks/browse/featured-playlists';
 
 function FeaturedPlaylistsPage(): JSX.Element {
