+++ src/js/pages/Browse/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { useLocation } from 'react-router-dom';
-import Navigation from '../../components/Navigation';
+import { Navigation } from '../../components';
 
 function BrowsePage(): JSX.Element {
   const { pathname } = useLocation();
